fix error database connection  and add user class

+++ backend/src/main/java/com/chattingweb/backend/UserRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
     List<User> findByFirstName(String firstName);
     List<User> findByAgeLessThan(Integer age);
-
 }
